Add millimeters to pixels converter, more

+++ LayoutEngine/CSS/CSSProperty.cs
@@ -2,7 +2,7 @@
 {
     class CSSProperty
     {
-        public double Value;
+        public float Value;
         public Unit Unit;
     }
 }
